backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
channels>=4.0
daphne>=4.0
stripe>=8.0
psycopg2-binary>=2.9
django-environ>=0.11
